Split view methods out of the Molecule class.
