+++ src/index.js
@@ -27,6 +27,7 @@ const server = http.createServer((request, response) => {
     case '/link.js':
     case '/molecule.js':
     case '/molecule-data.js':
+    case '/view/molecule.js':
     case '/page.js':
     case '/search.js':
     case '/test.js':
